bad_drivers/requirements.txt
django
djangorestframework
gunicorn
gevent
psycopg2-binary
djoser
djangorestframework_simplejwt